Update ringct.md
I didn't intend to change so much, but I found that we were explaining the same concepts two and three times.  Also, as a result, some rearrangements to the order of presentation were made.  I'm sure the repetition was largely an artifact of my original copy.  Anyway, I tried to strip out repetitive items, and generally streamline the information flow.  That said, I admit it's definitely not perfect; so, as usual, feel free to continue to chop and fix.

Note my [[comments]] where I recommended we exclude the commitment equations from the narration.  IMO, this is already super technical.  Having impressive real math on screen would certainly be awesome, but walking through the equation would be a mistake, in my opinion.