Make Alice -> Bob example more clear
I can't decide if it's too wordy.. but I think it's important to spoon-feed the point being made, so viewers don't have to figure out what is being implied. We're somehow trying to make this message both (a) palatable to total novices, and (b) useful/informative to reasonably-educated novices. That was my motivation for the rework.